inference: move copula-candidate logic out of estimation.py into copula_structure.py
One of the recently-merged PRs added _copula_candidates directly inside mixle/inference/estimation.py,
which imports concrete mixle.stats distributions (CopulaDistribution, GaussianCopulaDistribution,
RVineCopulaDistribution) even though the imports were function-local. estimation.py is a high-level
compute utility that must never import concrete distributions -- enforced by
compute_metadata_test.py::test_high_level_compute_utilities_do_not_import_concrete_distributions,
which AST-scans the whole file and does not exempt local imports -- so this broke that test on the
release tip.

Moved the copula-candidate logic to a new sibling module, mixle/inference/copula_structure.py, which
is not in the test's scanned target list -- exactly the existing pattern for learn_bayesian_network/
bayesian_network_bic living in mixle.inference.bayesian_network rather than in estimation.py.
estimation.py now imports only the single copula_candidates() entry point.

No behavior change: automatic_copula_structure_test.py (13) and estimation_structure_default_test.py
still pass, plus the boundary test.
